+++ myokit/_sim/opencl.py
@@ -123,8 +123,17 @@ class OpenCL(myokit.CModule):
         If ``formatted=True`` is set, a formatted version of the information is
         returned instead.
         """
+        if not OpenCL.supported():
+            if formatted:
+                return 'No OpenCL support detected.'
+            return OpenCLInfo()
+
         info = OpenCLInfo(OpenCL._get_instance().info())
-        return info.format() if formatted else info
+        if formatted:
+            if len(info.platforms) == 0:
+                return 'OpenCL drivers detected, but no devices found.'
+            return info.format()
+        return info
 
     @staticmethod
     def load_selection():
@@ -268,7 +277,7 @@ class OpenCLInfo(object):
 
     ``OpenCLInfo`` objects are created and returned by :class:`myokit.OpenCL`.
     """
-    def __init__(self, mcl_info):
+    def __init__(self, mcl_info=[]):
         # mcl_info is a python object returned by mcl_device_info (mcl.h)
         self.platforms = tuple([OpenCLPlatformInfo(x) for x in mcl_info])